requirements.txt
openai
langchain
faiss-cpu
chromadb      
openai 
tiktoken
python-dotenv
sentence-transformers
requests
bs4
rich
gradio
ruff